requirements.txt

fastapi==0.104.1
uvicorn==0.24.0
pyswisseph==2.10.3.2
pydantic==2.4.2
pytz==2023.3
python-multipart==0.0.6
typing-extensions==4.8.0
openai==1.51.2
python-dotenv==1.0.0
